streaming-prototype/requirements.txt
kafka-python==2.0.2
duckdb>=0.10.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
